+++ compiler/ghcjs/ghcjs.nix
@@ -9,7 +9,8 @@
 , ghc ? pkgs.buildPackages.ghc
 }:
 let
-    isGhcjs88 = builtins.compareVersions ghcjsVersion "8.8.0.0" > 0;
+    isGhcjs88 = builtins.compareVersions ghcjsVersion "8.8.0.0" >= 0;
+    isGhcjs810 = builtins.compareVersions ghcjsVersion "8.10.0.0" >= 0;
 
     project = pkgs.buildPackages.haskell-nix.ghcjsProject {
         src = ghcjsSrc;
@@ -104,10 +105,14 @@ let
           fi
         done
         } &
-        ${ if isGhcjs88
-          then
-            # Unsets NIX_CFLAGS_COMPILE so the osx version of iconv.h is not used by mistake
+        ${
+          # Unsets NIX_CFLAGS_COMPILE so the osx version of iconv.h is not used by mistake
+          if isGhcjs810 then ''
+            env -u NIX_CFLAGS_COMPILE PATH=$out/bin:$PATH \
+              $out/bin/ghcjs-boot --with-emsdk=${project.emsdk} --no-haddock \
+              || (echo failed > $TMP/done; false)
             ''
+          else if isGhcjs88 then ''
             env -u NIX_CFLAGS_COMPILE PATH=$out/bin:$PATH \
               $out/bin/ghcjs-boot -j1 --with-emsdk=${project.emsdk} --no-prof --no-haddock \
               || (echo failed > $TMP/done; false)