+++ lib/default.nix
@@ -334,6 +334,23 @@ in {
             else src.origSrc or src;  # If there is a subDir and origSrc (but no filter) use origSrc
   };
 
+  # Run evalModules passing the project function argument (m) as a module along with
+  # the the a projectType module (../modules/cabal-project.nix or ../modules/stack-project.nix).
+  # The resulting config is then passed to the project function's implementation.
+  evalProjectModule = projectType: m: f: f
+      (lib.evalModules {
+        modules = [m] ++ [
+          # Include ../modules/cabal-project.nix or ../modules/stack-project.nix
+          (import projectType)
+          # Pass the pkgs to the modules
+          ({ config, lib, ... }: {
+            _module.args = {
+              inherit pkgs;
+            };
+          })
+        ];
+      }).config;
+
   # Converts from a `compoent.depends` value to a library derivation.
   # In the case of sublibs the `depends` value should already be the derivation.
   dependToLib = d: d.components.library or d;