+++ test/index-state/default.nix
@@ -37,6 +37,11 @@ in
     '';
 
     meta.platforms = platforms.all;
+    # This test will need to be updated to use newer hackage index-state for it
+    # to work with GHC 9 and above.
+    # Does not work for GHCJS
+    meta.disabled = stdenv.hostPlatform.isGhcjs || compiler-nix-name == "ghc901";
+    
 
     passthru = {
       # Used for debugging with nix repl